% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabio.R
\name{fabio-package}
\alias{fabio-package}
\alias{fabio}
\title{Forestry and Agriculture Biomass Input-Output Tables}
\description{
In 2015 we started building physical supply-use and input-output tables covering global agriculture and forestry. 
The work is fully based on freely available data from FAOSTAT, IEA, EIA and UN Comtrade. FABIO now covers 191 countries, 
116 processes and 130 commodities for the years 1986-2013. The code will be released here in July 2018. An R package is in preparation [1].
}
\references{
[1] ...
}
\author{
Martin Bruckner, \email{martin.bruckner@wu.ac.at}
}
