% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FABIO.R
\name{fabio}
\alias{fabio}
\title{The main function to perform TWAS fine-mapping on binary outcomes}
\usage{
fabio(X, y, w_step = 6000, s_step = 20000)
}
\arguments{
\item{X}{Input predicted GReX matrix}

\item{y}{A vector of binary phenotypes of TWAS}

\item{w_step}{The number of warm-up steps in MCMC, default = 6000}

\item{s_step}{The number of sampling steps in MCMC, default = 20000}
}
\value{
PIP summary table
}
\description{
The main function to perform TWAS fine-mapping on binary outcomes
}
\examples{
# example.rda can be downloaded from: 
# https://www.dropbox.com/scl/fi/hhdjl7ep881cv8fx2hbdk/example.rda?rlkey=zanf4tyeqar49lhpu0efbpipf&dl=0

load('/path/to/example.rda')
X <- inputs[[1]]
y <- inputs[[2]]
output <- fabio(X, y, 100, 1000)
}
