% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabio_footprint.R
\name{fabio_footprint}
\alias{fabio_footprint}
\title{fabio footprint}
\usage{
fabio_footprint(wd, years, regions_select = "all")
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Footprint analysis based on fabio model.
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
