% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_0_read_raw_FAO_data.R
\name{FAO_MRIO_0_read_raw_FAO_data}
\alias{FAO_MRIO_0_read_raw_FAO_data}
\title{FAO_MRIO_0_read_raw_FAO_data}
\usage{
FAO_MRIO_0_read_raw_FAO_data(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
TBA
}
\examples{
\dontrun{

df <- FAO_MRIO_0_read_raw_FAO_data()

}
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
