% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_1b_data_preparation.R
\name{FAO_MRIO_1b_data_preparation}
\alias{FAO_MRIO_1b_data_preparation}
\title{FAO_MRIO_1b_data_preparation}
\usage{
FAO_MRIO_1b_data_preparation(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 1b: prepare CBS and BTD data sets
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
