% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_1c_estimate_missing_CBS.R
\name{FAO_MRIO_1c_estimate_missing_CBS}
\alias{FAO_MRIO_1c_estimate_missing_CBS}
\title{FAO_MRIO_1c_estimate_missing_CBS}
\usage{
FAO_MRIO_1c_estimate_missing_CBS(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 1c: estimate missing CBS
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
