% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_1e_start_values_BTD.R
\name{FAO_MRIO_1e_start_values_BTD}
\alias{FAO_MRIO_1e_start_values_BTD}
\title{FAO_MRIO_1e_start_values_BTD}
\usage{
FAO_MRIO_1e_start_values_BTD(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 1e: define start values for BTD
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
