% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_1f_BTD_balancing.R
\name{FAO_MRIO_1f_BTD_balancing}
\alias{FAO_MRIO_1f_BTD_balancing}
\title{FAO_MRIO_1f_BTD_balancing}
\usage{
FAO_MRIO_1f_BTD_balancing(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 1f: balance BTD_original + BTD_est using RAS
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
