% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_2a_supply.R
\name{FAO_MRIO_2a_supply}
\alias{FAO_MRIO_2a_supply}
\title{FAO_MRIO_2a_supply}
\usage{
FAO_MRIO_2a_supply(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 2a: build supply tables
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
