% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_2b_use.R
\name{FAO_MRIO_2b_use}
\alias{FAO_MRIO_2b_use}
\title{FAO_MRIO_2b_use}
\usage{
FAO_MRIO_2b_use(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 2b: build use tables
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
