% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_2c_trade_linking_SUP.R
\name{FAO_MRIO_2c_trade_linking_SUP}
\alias{FAO_MRIO_2c_trade_linking_SUP}
\title{FAO_MRIO_1b_data_preparation}
\usage{
FAO_MRIO_2c_trade_linking_SUP(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 2c: trade linking of supply tables
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
