% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_2d_trade_linking_USE.R
\name{FAO_MRIO_2d_trade_linking_USE}
\alias{FAO_MRIO_2d_trade_linking_USE}
\title{FAO_MRIO_2d_trade_linking_USE}
\usage{
FAO_MRIO_2d_trade_linking_USE(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 2d: trade linking of use tables
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
