% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_3a_mriot.R
\name{FAO_MRIO_3a_mriot}
\alias{FAO_MRIO_3a_mriot}
\title{FAO_MRIO_3a_mriot}
\usage{
FAO_MRIO_3a_mriot(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 3a: construct multi-regional input-output tables
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
