% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_3b_A.R
\name{FAO_MRIO_3b_A}
\alias{FAO_MRIO_3b_A}
\title{FAO_MRIO_3b_A}
\usage{
FAO_MRIO_3b_A(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 3b: derive A matrix
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
