% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_3c_L_series.R
\name{FAO_MRIO_3c_L_series}
\alias{FAO_MRIO_3c_L_series}
\title{FAO_MRIO_3c_L_series}
\usage{
FAO_MRIO_3c_L_series(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 3c: derive Leontief inverse matrix via power series
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
