% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_4a_simple_A_footprints.R
\name{FAO_MRIO_4a_simple_A_footprints}
\alias{FAO_MRIO_4a_simple_A_footprints}
\title{FAO_MRIO_4a_simple_A_footprints}
\usage{
FAO_MRIO_4a_simple_A_footprints(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 4a: calculate simple A footprints
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
