% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAO_MRIO_4c_footprints_AB.R
\name{FAO_MRIO_4c_footprints_AB}
\alias{FAO_MRIO_4c_footprints_AB}
\title{FAO_MRIO_4c_footprints_AB}
\usage{
FAO_MRIO_4c_footprints_AB(...)
}
\arguments{
\item{...}{some argument}
}
\value{
returns some data
}
\description{
Build MRIO table based on Faostat commodity balance sheets and trade data part 4c: calculate footprints AB (?)
}
\author{
Sebastian Luckeneder, \email{sebastian.luckeneder@wu.ac.at}
}
